Finalize update to directus for ModalPanziGong
